add check course and session owner, add check user visibility to teacher.

+++ src/main/java/com/app/manager/service/implementClass/CourseServiceImp.java
@@ -1,10 +1,11 @@
 package com.app.manager.service.implementClass;
 
+import com.app.manager.context.repository.CourseRepository;
+import com.app.manager.context.repository.UserRepository;
 import com.app.manager.context.specification.CourseSpecification;
 import com.app.manager.entity.Course;
 import com.app.manager.model.payload.CourseModel;
 import com.app.manager.model.returnResult.DatabaseQueryResult;
-import com.app.manager.context.repository.CourseRepository;
 import com.app.manager.service.interfaceClass.CourseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -12,8 +13,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -21,6 +20,9 @@ public class CourseServiceImp implements CourseService {
     @SuppressWarnings("SpringJavaAutowiredFieldsWarningInspection")
     @Autowired
     CourseRepository courseRepository;
+    @SuppressWarnings("SpringJavaAutowiredFieldsWarningInspection")
+    @Autowired
+    UserRepository userRepository;
 
     @Override
     public Page<CourseModel> findAll(CourseSpecification courseSpecification, Pageable pageable) {
@@ -70,8 +72,14 @@ public class CourseServiceImp implements CourseService {
     }
 
     @Override
-    public DatabaseQueryResult update(CourseModel courseModel, String id) {
+    public DatabaseQueryResult update(CourseModel courseModel,
+                                      String id, String currentUsername) {
         try {
+            var teacher = userRepository.findByUsername(currentUsername);
+            if(teacher.isEmpty())
+                return new DatabaseQueryResult(false, "Teacher not found",
+                        HttpStatus.NOT_FOUND, "");
+
             var c = courseRepository.findById(id);
             if(c.isEmpty()){
                 return new DatabaseQueryResult(false,
@@ -79,6 +87,11 @@ public class CourseServiceImp implements CourseService {
             }
 
             var course  = c.get();
+
+            if(!course.getUserid().equals(teacher.get().getId()))
+                return new DatabaseQueryResult(false, "Not your course",
+                        HttpStatus.BAD_REQUEST, "");
+
             course.setCoursecategoryid(courseModel.getCoursecategoryid());
             course.setCreatedat(courseModel.getCreatedat());
             course.setDescription(courseModel.getDescription());
@@ -96,13 +109,22 @@ public class CourseServiceImp implements CourseService {
     }
 
     @Override
-    public DatabaseQueryResult delete(String id) {
+    public DatabaseQueryResult delete(String id, String currentUsername) {
         try {
+            var teacher = userRepository.findByUsername(currentUsername);
+            if(teacher.isEmpty())
+                return new DatabaseQueryResult(false, "Teacher not found",
+                        HttpStatus.NOT_FOUND, "");
+
             var course = courseRepository.findById(id);
             if(course.isEmpty()){
                 return new DatabaseQueryResult(false,
                         "delete course failed", HttpStatus.NOT_FOUND, "");
             }
+            if(!course.get().getUserid().equals(teacher.get().getId()))
+                return new DatabaseQueryResult(false, "Not your course",
+                        HttpStatus.BAD_REQUEST, "");
+
             courseRepository.delete(course.get());
             return new DatabaseQueryResult(true,
                     "delete course success", HttpStatus.OK, "");
